+++ tsconfig.json
@@ -18,7 +18,10 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "noUncheckedSideEffectImports": true
+    "noUncheckedSideEffectImports": true,
+
+    "jsx": "react",
+    "jsxFactory": "Arwaal.createElement"
   },
   "include": ["src"]
 }
